refactor(Parser) : Parser의 int 변환을 long 변환으로 수정

+++ src/main/java/calculator/view/OutputView.java
@@ -2,7 +2,7 @@ package calculator.view;
 
 public class OutputView {
 
-    public void printResult(int result) {
+    public void printResult(long result) {
         System.out.println("결과 : " + result);
     }
 }
